+++ CarRentalManagementSystem-ejb/src/java/ejb/session/stateless/ReservedSessionBeanLocal.java
@@ -6,6 +6,7 @@
 package ejb.session.stateless;
 
 import entity.Reserved;
+import java.math.BigDecimal;
 import java.util.List;
 import javax.ejb.Local;
 
@@ -28,7 +29,7 @@ public interface ReservedSessionBeanLocal {
 
     List<Reserved> viewAllReservationOfCustomer(long customerId);
 
-    Long CancelReservation(long reservationId);
+    BigDecimal CancelReservation(long reservationId);
 
 
 }
